[runtime] Add __attribute__(format) to glib string formatting functions, fix -Wformat warnings/errors exposed by the change. (mono/mono#11611)
[runtime] Add __attribute__(format) to glib string formatting functions, fix -Wformat warnings/errors exposed by the change.



Commit migrated from https://github.com/mono/mono/commit/d768d8a571413d475ce98acad100d04b4147896d

+++ src/mono/mono/mini/interp/mintops.c
@@ -58,7 +58,7 @@ mono_interp_dis_mintop(const guint16 *base, const guint16 *ip)
 	guint32 token;
 	int target;
 
-	g_string_append_printf (str, "IL_%04x: %-10s", ip - base, mono_interp_opname [*ip]);
+	g_string_append_printf (str, "IL_%04x: %-10s", (int)(ip - base), mono_interp_opname [*ip]);
 	switch (mono_interp_opargtype [*ip]) {
 	case MintOpNoArgs:
 		break;
@@ -115,7 +115,7 @@ mono_interp_dis_mintop(const guint16 *base, const guint16 *ip)
 			if (i > 0)
 				g_string_append_printf (str, ", ");
 			offset = (gint32)READ32 (p);
-			g_string_append_printf (str, "IL_%04x", p + offset - base);
+			g_string_append_printf (str, "IL_%04x", (int)(p + offset - base));
 			p += 2;
 		}
 		g_string_append_printf (str, ")");
